requirements.txt
discord.py==2.3.1
requests==2.27.1
python-dotenv==0.20.0
pytz
openai==0.27.8
PyYAML==6.0
dacite==1.6.0